backend/requirements.txt
fastapi
uvicorn
pdf2image
google-generativeai
networkx
websockets
python-multipart
python-dotenv
requests
